chore(websocket): Fix cookie and hub(room) behavior
Created option to join multiple clients to the same hub and exchange
actions. Check WSS repository for server side implementation.

Thinking, how to scope users according their organizations and views.
I don't need to sync actions if 2 users are in different views.
Current idea is to create hub for every open view per organization.
For example, user A from Org. A opens an Order view. I create hub
scoped for that organization and view. If user B from Org. A joins
the same Order view, i do add it to the same existing Hub. From this
point all actions could be synced for booth users.
Other issue is that Hub broadcasts/reflects actions to the sender user
itself. I should send messages only to the users in Hub which are not
origins of those action messages.
Will continue to explore.