feat(launcher): fetch list of the applications fix #115
Totally dirty implementation. Naming is wrong and organization as
well. Because i am retriecing a list of the applications it is not
clear to me should i store that state in Launcher feature or
should i keep it in Applications feature. I tend to think, that every
state piece should be as close as possible to the feature which is
using it. Decoupling. I will play more with this and see how it goes.